More converters.  Realised they need restructuring to deal with e.g. Enums.

+++ src/main/java/com/mcleodmoores/excel4j/typeconvert/converters/DoubleXLNumberTypeConverter.java
@@ -22,6 +22,6 @@ public final class DoubleXLNumberTypeConverter extends AbstractTypeConverter {
 
   @Override
   public Object toJavaObject(final XLValue from) {
-    return ((XLNumber) from).getValue();
+    return (double) ((XLNumber) from).getValue();
   }
 }
